+++ DataObjects/TournamentType.cs
@@ -16,6 +16,6 @@ namespace DataObjects
         public int? NumberOfPlayers { get; set; }
         public List<Reward> Reward { get; set; }
         public double? EntryFee { get; set; }
-        public int? OpenForTime { get; set; }
+        public double? OpenForTime { get; set; }
     }
 }
